+++ data/data-init.ts
@@ -5,12 +5,12 @@ export const menuGroups = [
         title: AdminMenuGroup.FAQ,
         items: [
             {
-                label: AdminMenuLabel.FAQ_CATEGORY,
-                href: AdminMenuPath.FAQ_CATEGORY,
-            },
-            {
                 label: AdminMenuLabel.FAQ_LIST,
                 href: AdminMenuPath.FAQ_LIST,
+            },
+            {
+                label: AdminMenuLabel.FAQ_CATEGORY,
+                href: AdminMenuPath.FAQ_CATEGORY,
             },
             {
                 label: AdminMenuLabel.FAQ_REGISTER,