+++ src/app/pub/faq/page.tsx
@@ -5,7 +5,6 @@ import useSWR from 'swr';
 import axiosInstance from '../../../../libs/axios';
 import type { Request as ApiRequest } from '../../../../types/Common';
 import type { MajorCategoryData } from '../../../../types/Faq';
-import { useEffect } from 'react';
 
 const fetcher = (payload: ApiRequest) => axiosInstance.post('/api/backend', payload).then((res) => res.data.result);
 const Page = () => {
@@ -23,10 +22,6 @@ const Page = () => {
             fallbackData: [],
         },
     );
-
-    useEffect(() => {
-        console.log(faqData);
-    }, [faqData]);
 
     return <Faq faqData={faqData} />;
 };